NCBC-972: High latency on async set with persistence to master
Motivation
----------
When doing an aynchronous observe and a rebalance occurs, make the
KeyObserver await while acquiring a server.

Modifications
-------------
Adds a Task.Delay in place of Thread.Sleep while trying to get a server to
execute the operation on.

Results
-------
Instead of the blocking the thread, the task will be awaited.

Change-Id: I665354dc2da0423b904e13a8a7da786b570cdfd9
Reviewed-on: http://review.couchbase.org/54231
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>