NCBC-969: Make timeout parameter default to 75s for N1QL queries
Motivation
----------
WIthout explicitly providing a value for the timeout parameter in a N1QL
query, the query could potentially run forever. This commit adds a default
timeout of 75s if the the user doesn't provide one; this mirrors the
client HTTP request timeout.

Modifications
-------------
Changed the QueryRequest class to use a default value of 75000(ms) for the
server timeout if the server timeout is not provided (0), when the query request
is generated.

Results
-------
All N1QL queries requests will use a default server and client timeout of
75s if none is provided explicitly. These values can be modified by the
ClientConfiguration.QueryDefaultTimeout value for the client HTTP request
and via the QueryRequest.Timeout(timeout) method for the server timeout.

Change-Id: I45cf12709c99f146ae59161e5699902a96945e6a
Reviewed-on: http://review.couchbase.org/54313
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>