NCBC-1011: Make ContinueOnAnyContext not capture current context
Motivation
-------------
Task.ConfigureAwait(false) does not modify the task, but returns a
ConfiguredTaskAwaitable which wraps the task in an awaiter that does not
capture the current context. Currently, the ContinueOnAnyContext extension
method is a no-op, and all tasks are capturing the current context.

Modifications
-------------
Change TaskExtensions.ContinueOnAnyContext to return
ConfiguredTaskAwaitable instead of Task.

Results
-------
Tasks will not capture the current context.

Change-Id: I6e1e2a021eb3b18ccfeaa70ab60331994bb014ff
Reviewed-on: http://review.couchbase.org/55034
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>