NCBC-752: Add support for Geo-Spatial Views
Motivation
----------
Couchbase Server supports multi-dimensional spatial indexes which can be
created by using views. A spatial index contains geometry data that can be
queried to determine whether recorded geometries exist within a given
range.

Modifications
-------------
A new interface and implementation called ISpatialViewQuery has been
defined. It's purpose is to create a URI which will then be executed
against the Server. A new interface was introduced, called IViewQueryable,
which allows the IViewClient to accept any implementation and execute the
request, mapping the results back to an IViewResult impl.

Results
-------
You can now use the Couchbase .NET client to execute geo-spatial view
requests if the target cluster version if Couchbase Server 4.0 or higher.
If the server does not support the feature, a NotSupportedException will
be returned in the IViewResult.Exception field.

Change-Id: I9fb46e78f0c020714adb0b671657803b852b914e
Reviewed-on: http://review.couchbase.org/54804
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>