NCBC-983: Improve handling of temporary network instability
Motivation
----------
This commit changes the behavior of how the SDK determines that a remote
host (node/couchbase server) is down and the reconnect policy to check if
it has gone back online.

Modifications
------------
The logic determining whether a node is down was changed in the Server
class (representing a remote cluster node). IServer.TakeOffline was renamed
and modified so that the count of errors within a certain interval was
used as a determining factor of whether a node is down. If a operation
times out at the transport level, this most certianly is a remote host
either going offline or a service crash on that remote host, so this will
immediately put the node in a "down" state. A new exception type
RemoteHostTimeoutException was created for this situation. IServer.IsDead was
renamed to IServer.IsDown to better reflect the behavior.

Results
-------
Previous to this commit, the SDK would handle any network issue by putting
the node into a "down" state. This commit changes that behavior so that a
certain number of errors (IOErrorThreshold) within a certain time period
(IOErrorInterval) must occur before the node will be put into this state
or if an operation times out in mid-flight indicating the remote host has
gone offline. In most cases, if an IO error occurs, the connection will
immediately be recreated. Note, in this case the operation will recieve a
ResponseStatus of TransportFailure and may or may not be retried depending
upon the type of operation and if it's a mutation, whether it has a CAS
associated with it.

Change-Id: Icc4f5e9ab229d5993c14ac174f9aac81b87a0a47
Reviewed-on: http://review.couchbase.org/54995
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>