NCBC-965: Insert with dollars would fail, add test
Motivation
----------
INSERT statements with values containing dollar sign or & would at some
point fail. With the recent changes this has been corrected, but we
want to add a test with such values to prevent regressions.

Modifications
-------------
Added an integration test for this issue to prevent regressions.

Result
------
It is verified that INSERT statements accept values with dollar or &.

Change-Id: I76cbfbf0d1e6849cac15c4f369fd3c02b6685f0e
Reviewed-on: http://review.couchbase.org/55084
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>