NCBC-950: SSL/TLS support for N1QL
Motivation
----------
This commit enables SSL/TLS support for N1QL query request, which means
that the data sent to and from Couchbase server will be encrypted.

Modifications
-------------
Refactored the QueryClient to use input from configuration when generated
the URI for the query request; this impacts both ViewQuery and
SpatialViewQuery, so they were updated as well. If UseSsl is set to true, the URI
generated will use the HTTPS protocol and port 18093. Also, refactored URI
generation and created a dedicated UriUtil utility class for generating
URI's.

Results
-------
If ClientConfiguration.UseSsl is true and a certificate is installed on
the app server, the SDK will ensure that data sent to and from the
Couchbase server will be encrypted.

Change-Id: I2a0263c8a32217719acb41920e6209845832984b
Reviewed-on: http://review.couchbase.org/55117
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>