NCBC-985: Auth error during rebalance with Memcached bucket and SASL auth/password
Motivation
----------
This commit fixes a bug where a Memcached bucket would fail to
authenticate after a rebalance.

Modifications
-------------
The ConfigThreadState was updated so that the password stored with the
initial BucketConfig would be passed to the latest BucketConfig when a
clustermap change happens.

Results
-------
If a password protected Memcached bucket is opened and rebalanced
authenication will be successfully transferred when the clustermap
updates.

Change-Id: I1e6b537744e0a99a1e0a8d80abc59ffee5cf13e0
Reviewed-on: http://review.couchbase.org/55150
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>