NCBC-986: Bucket name is missing from URI created by ViewQuery
Motivation
----------
View queries require the bucket be included within the path of URI.

Modifications
-------------
Modified ViewQuery so that it will use the bucket name if it exists or
just rely on the base URI if it has a bucket.

Result
------
ViewQuery will ensure that the URI is properly constructed with the
bucketname included.

Change-Id: I82eda47cabe89da1c10771424f48e5efc31c4165
Reviewed-on: http://review.couchbase.org/55276
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>