NCBC-990: Spatial View queries do not map the geometry field in the results
Motivation
----------
Geo Spatial Views optionally emit the geometry structure per the GEOJson
specification: http://geojson.org/geojson-spec.html.

Modifications
-------------
The ViewRow class now has an additional field Geometry which maps to the
geometry if it is emitted by a GEO Spatial View.

Results
-------
If a Geo spatial View emits geometry, it will be mapped to the ViewRow; if
not, the property will be null.

Change-Id: I899510e11ae7b1572edb6e79a5012ea00817ac87
Reviewed-on: http://review.couchbase.org/55351
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>