NCBC-996: Add support for provisioning a cluster to ClusterManager
Motivation
----------
Adds additional support for programatically provision and configuring a
cluster.

Modifications
-------------
Updates ClusterManager to support adding services (data, query, or index),
configuring memory, adding an admin and adding a new node to an existing
cluster. Also, introduces a new convience class called ClusterProvisioner
that makes it easier to set up a cluster. Note that this class is
"experimental" and subject to change in future releases.

Results
-------
It's now possible to provision a cluster from VM's (for example) and
configure the cluster programatically.

Change-Id: I6e53cb232e8eae5be785c2ee3d15e51a3fdb9b17
Reviewed-on: http://review.couchbase.org/55533
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>