NCBC-992: Query nodes that are unreachable are not taken out of rotation
Motivation
----------
If a request is mapped to a query node that is no longer onlone or cannot
process the request, that node should be taken offline and the request
retried on another node if available.

Modifications
-------------
The CouchbaseRequestExecutor was modified so that it retried query
requests if the QueryResponse was not Fatal or Errors. Errors indicates
that the query maybe malformed for example. Server now uses a query
request for a heartbeat check if the node is a query node.

Results
-------
When a query node is offline or unresponsive, the client will mark it as
dead. If the request that failed (triggering the dead status) is
retryiable, it wil be retried on a different node. A hertbeat check will
then be started and if successful, then the node will be put back into
rotation.

Change-Id: Ic823f6e4be61b11274d9fdc12923f62930f38762
Reviewed-on: http://review.couchbase.org/55681
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>