NCBC-1014: Refactor unit tests
Motivation
----------
It would be nice if  the UT's and integration tests could be run in an
environment by only changing a single bootstrap setting in the App.Config.

Modifications
-------------
Made all unit tests depend up the ClientConfigUtil.GetConfiguration()
method which uses the bootstrapUrl AppSetting value in the App.Config.
Fixed and refactored other tests that were no longer working.

Results
-------
To change the environment you wish to run the tests on, change the
bootstrapUrl in the AppSettings section of the App.Config.

Change-Id: Iba466e676a408d8341bd2c344d9af6621acec04c
Reviewed-on: http://review.couchbase.org/55752
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>