NCBC-1006: only retry when query response is HTTP 400 level
Motivation
----------
For example, if a primary index is not defined, the client will receive a HTTP 404; the SD
will then attempt to retry and would loop indefinitely. This commit allows
retries only on non-400 errors.

Modifications
-------------
Added a Lifespan struct representing a duration of time for an operation. This
can be reused for other purposes. If the query takes more than
ClientConfiguration.QueryRequestTimeout and attempts a retry, it will terminate
the loop. Added logic to determine if HTTP 404 is returned, if not retries are allowed.

Results
-------
A HTTP 400 level error will fail-fast; other failures will be retried up until
ClientConfiguration.QueryRequestTimeout is reached.

Change-Id: I3db417ee201b127844a74b44afd726195c9b0744
Reviewed-on: http://review.couchbase.org/55797
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>