NCBC-1009: Fix ObjectDisposedException if Finalizer runs
Motivation
----------
If finalizer runs (debug build only) after object has disposed, an ODE will
be thrown. This fix will suppress the exception and try to log it. This should
only impact test environments.

Modifications
-------------
Try/Catch added to ConnectionPool.Finalize - it will try to log the exception
if it cannot, it will suppress it.

Results
-------
ObjectDisposedException will not be thrown if the client is compiled with
DEBUG flags and the finalizer is run during a dispose.

Change-Id: I3cbf1dc857df96c30f54e353d384790bc02735d3
Reviewed-on: http://review.couchbase.org/55944
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>