NCBC-1020: Exists should return false if key is LogicalDeleted
Motivation
----------
Exists function shouldn't need to wait until after a Remove is persisted
to disk before it returns the expected result of false.

Modifications
-------------
Return false for KeyState.LogicalDeleted in addition to KeyState.NotFound.

To support unit testing, added an internal overload to the CouchbaseBucket
constructor that accepts an IRequestExecuter.  Also modified the
_requestExecuter field to store it as an IRequestExecuter instead of a
CouchbaseRequestExecuter.