NCBC-1030: Query requests are sent to rebalanced out query node
Motivation
----------
If a query node is swapped out and the bootstrap node is a different node
and the query node is also not a data node, the cluster map updates will
not be recieved by the client, thus the client will continue to use an old
config. This patch fixes that bug by forcing a config update check if a
certain exception is thrown.

Modifications
-------------
Add the UpdateConfig method to the IRequestExecutor interface and an
implementation in the base class. Add additional logic to Server class so
that references do not live on after being disposed because of a leaky
timer.

Result
------
If a non-data node recieves a remote error (the node is no longer part of
the cluster), it will force a config check and update the cluster map if
it's different from the current config.

Change-Id: Icdad5140f8111fbc34378972e6834aebed9346a1
Reviewed-on: http://review.couchbase.org/56714
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>