NCBC-1039: Force QueryClient to use HTTP Auth for N1QL queries
Motivation
----------
Work around for MB-16964. Forces a N1QL query to use HTTP Auth headers as
opposed to the "creds" array which is not working correctly for HTTP POST.

Modifications
-------------
Added Auth headers to HttpClient and use the bucket name and password as
credentials.

Result
------
The QueryRequestAddCrendentials(..) method does not currenly work; this is
a work around.

Change-Id: I0ae3a2eb9d552ee396322be9bf6826b61441ba0f
Reviewed-on: http://review.couchbase.org/57474
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>