NCBC-1042: Add support for max_parallelism to N1QL query request
Motivation
----------
Specifies the maximum parallelism for the query on the remote server. There
is also a server wide max_parallelism parameter which defaults to 1. If a
request includes max_parallelism, it will be capped by the server
max_parallelism. If a request does not include max_parallelism, the server
wide max_parallelism will be used.

Modifications
-------------
Add MaxServerParallelism method to QueryRequest; the "Server" in the name
is to indicate the parallelism configuration is for server-side
parallelism and not client-side parallalism related to async/await, TPL or
multi-threading.

Results
-------
Server-side parameter for max_parallelism is supported by the QueryRequest
class.

Change-Id: Ie84e3a3a39f27caad447b2991f301d93922c534b
Reviewed-on: http://review.couchbase.org/57799
Well-Formed: buildbot <build@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>