NCBC-1039: Part 2 - Force QueryClient to use HTTP Auth for N1QL queries
Motivation
----------
Part 2 of a work around for MB-16964. Forces a N1QL query to use HTTP Auth
headers as opposed to the "creds" array which is not working correctly for
HTTP POST. This commit is specifically for non-async queries. Part 1 can
be found here: 3e19deec20f230fbbe2bcdf395956b6e3432e16c.

Modifications
-------------
Added Auth headers to HttpClient and use the bucket name and password as
credentials.

Result
------
The QueryRequestAddCrendentials(..) method does not currenly work; this is
a work around.

Change-Id: I9bfc237066d8105249a6bbcd83b6a20e9af4d7d8
Reviewed-on: http://review.couchbase.org/57820
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>