NCBC-1045: Add support for pluggable IO service implementations
Motivation
----------
Consumers may want to roll their own IO services based upon different
performance profiles or use-cases; this allows the IO
service and connection to be pluggable.

Modifications
-------------
Adds a factory method to the ClientConfiguration and ClientConfigSection
that allows the IO service/strategy to be injected.

Results
-------
The IO Service and connection layer is now extensible/pluggable.

Change-Id: Ic513b80ac874afe1e6cc061f4095874d0061aaf3
Reviewed-on: http://review.couchbase.org/57968
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>