NCBC-1046: Make MemcachedConfigContext use NodesAdapter for comparisons
Motivation
----------
The MemcachedConfigContext was using the Nodes list as a comparison to see
if the client needed a full reconfiguration; the NodesAdapter list is the
authoritative list and some config changes may be missed.

Modifications
-------------
Updated the MemcachedConfigContext so that it compares the NodesAdapter
list rather than the Nodes list when a config update is detected.

Result
------
The client will use the most up-to-date cluster configuration.

Change-Id: I799d723f6c81b6d4d456efdcbc43ea207ab1f10d
Reviewed-on: http://review.couchbase.org/58029
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>