NCBC-1049: Rename IOStrategy to IOService and other name refactoring
Motivation
----------
Improve the naming of core IO namespaces and classes.

Modifications
-------------
The namespace Couchbase.IO.Strategies was renamed to
Couchbase.IO.Services. IOStrategy and related classes were renamed to
IIOServices. All variable names and namespaces were renamed appropriately.

Results
-------
The IO namespace interfaces and classes are now named in a manner
consistant with their usage and behavior. This is important as some of
the Types in this namespace may become public in future releases so that
the SDK is more extensibility.

Change-Id: Idb7d17f6f8f17b4379ce8a85b35cad994c5a5aa0
Reviewed-on: http://review.couchbase.org/58144
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>