NCBC-1047: log current SendTimeout if request times out at transport layer
Motivation
----------
Previously the client would log the default value of the SendTimeout
setting but not the actual value. This made debugging difficult.

Modifications
-------------
The current SendTimeout value is now logged on SslConnection and
Connection classes. The message was refactored and now is on the
ExceptionUtil class.

Results
-------
If a "send" request times out, then the client will log the SendTimeout
value from the configuration.

Change-Id: I00a63f2256365ecea4b3d33e85d699f0213f89a0
Reviewed-on: http://review.couchbase.org/58153
Reviewed-by: Brett Lawson <brett19@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>