NCBC-1076: Use default server configuration when there is no bucket config
Motivation
----------

If in the web.config <couchbase> section we only have a <servers> section,
but no <buckets> section (because we don't want to statically configure
any buckets), then when we try to open a bucket using its name, the client
tries to use localhost:8091 instead of the server configured <servers>.
See
http://stackoverflow.com/questions/34183342/why-can-opening-a-couchbase-bucket-be-very-slow
for more details.

Modifications
-------------
Use parent configuration settings when creating a BucketConfiguration in
ConfigProviderBase.

Results
-------
When we get the configuration for a bucket, the server
configuration is getting used even if there is no <buckets> section.

Change-Id: I3e8f75e0a8b54c396806923b3e6cf8d376b954c3
Reviewed-on: http://review.couchbase.org/58348
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>