NCBC-1057: Don't fail fast when service is unavailable
Motivation
----------
If the cluster does not contain a candidate node for a client request
(data, index, query) instead of failing fast, the client should return
back a failed response with a ServiceUnavailableException populating the
IRequest.Exception field.

Modifications
-------------
Modified the appropriate methods for each request type to return the
ServiceUnavailableException rather than throw the exception.

Result
------
In certain situations such a node swap, returning the exception is preferable
to throwing an exception since the failure likely temporary and the node will
soon be back online.

Change-Id: I2d39d4543cd109d93b97ef90b39ba666fe99c762
Reviewed-on: http://review.couchbase.org/58504
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>