NCBC-998: Support for Sub-Document API - part 1
Motivation
----------
This is the first of several commits introducing the Developer Preview for
the Sub-document API in the Couchbase .NET SDK. Disclaimer: this API is
not tested nor supported other than for testing and/or exploratory
purposes. In subsequent releases the remainder of the API will be released
and eventually tested for GA release.

Modifications
-------------
Two new methods were added to the IBUcket interface and implemented in
CouchbaseBucket (MemcachedBucket's implementation is not supported):
MutateIn and LookupIn. MutateIn returns the builder API for mutation
operations and LookupIn returns the builder API for lookup operations. A
new return type called DocumentFrament{T} was introduced which represents
the result of a multiple sub-document operations with special helper
methods for access the value of the operation if it exists and to check on
the status of an operation. Finally, the low-level Memcached extensions
for multi-commands were added.

Results
-------
With this commit, sub document API is introduced for multi-commands. It is
a Developer Preview and has not gone through the full range of testing nor
is the interface complete; it's not suggested that the Sub-document API be
used in production scenarios.

Change-Id: I66a3e9160e124f29ec3d63262c37bfff3d748b90
Reviewed-on: http://review.couchbase.org/58717
Reviewed-by: Matt Ingenthron <matt@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>