NCBC-1074: Prevent MVC deadlock on async view queries
Motivation
----------
When working within MVC actions, executing an asynchronous view query and
then waiting for it to complete synchronously causes a deadlock.

Modifications
-------------
CouchbaseRequestExecutor for aysnchronous view queries needed an
additional call to ContinueOnAnyContext before an await.

Also created an integration test which successfully detects this problem
if it occurs in the future.

Results
-------
MVC actions no longer deadlock in this scenario.

Change-Id: I4e16062487e0b794dac413efc81e09341b96a809
Reviewed-on: http://review.couchbase.org/59320
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>