NCBC-1083: PoolConfiguration still uses default settings when overridden
Motivation
----------
If PoolConfiguration settings are set on the ClientConfiguration, they
must override the default settings if they have not changed.

Modifications
-------------
Modified ClientConfiguration to properly change these properties when they
have been modified.

Result
------
If the PoolConfiguration settings are changed on the ClientCOnfiguration,
then they will override the default settings if they have not changed.

Change-Id: I650d072ca5d6101f02dbb7968a5ad9a9e00aa213
Reviewed-on: http://review.couchbase.org/60071
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>