NCBC-981: When FQDN is defined for Couchbase instance SSL fail
Motivation
----------
When SslStream.AuthenticateAsClient is called, the targetHost must match
the name in the certificate. Previously the ipaddress was passed in as the
targetHost and if the cluster was configured with a hostname, the
validation would fail. This patch uses the hostname of the pools uri
(bootstrap) for each node as the targetHost.

Modifications
-------------
Add clone method to PoolConfiguration and a Uri property to hold each
node's specific bootstrap URI. Update the SslConnection class to use the
hostname as the targetHost when validating the certificate. Make
SslConnection.ServerCertificateValidationCallback check that
sslPoliciesErrors is NONE.

Results
-------
The hostname (or the IP address if it is in the URI) of the bootstrap URI or a candidate URI is
now used to validate against the certificate instead of expliclity using
the IP address.

Change-Id: I6d4200b152821aab0f576eb560ac5c4da3993a18
Reviewed-on: http://review.couchbase.org/60103
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>