NCBC-1086: GetAndLock not returning Locked status
Motivation
----------
Fixes a bug where GetAndLock will timeout on retries instead of returning
the correct response that the key has been locked.

Modifications
-------------
If a key is locked, the server response for a GETL will be a temp fail
(0x0086) and the body will contain "LOCK_ERROR". We use this information
to immediately return back to the caller without retrying the operation.

Result
------
If a key is locked, GetAndLock will fail with a TemporaryFailure and the
OperationResult.Message will contain the text "LOCK_ERROR".

Change-Id: Ic0888c22012aeddcdcd0afe3d5951020339223f2
Reviewed-on: http://review.couchbase.org/60440
Reviewed-by: Mark Nunberg <mark.nunberg@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>