NCBC-1089: Await callbacks if Task throws exception
Motivation
----------
In the exception handling of many methods, the callback is invoked
synchronously. This changes the behavior to await on the callback instead.
It also add sets ConfigureAwaiter to false.

Modifications
-------------
Change synchronous invocation to async by awaiting the callback. Set
ConfigureAwaiter to false so that task resumes on the current
synchronization context and not the one which triggered it. Since awaiting
in catch and finalize blocks is not allowed with earlier compilers, we
changed the logic to await after the critical blocks if the
ExceptionDispatchInfo is not null (thus an exception was handled).

Change-Id: I1aa6a71ef4f496f0e4acdfceb1d8ac264f3a2595
Reviewed-on: http://review.couchbase.org/60670
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>