Lambda navigation for sub document paths
Motivation
----------
Provide a simplier, syntax checked, intellisense supported method for
building subdocument paths.

Modifications
-------------
Made ILookupInBuilder and ILookupInBuilder generic based on the root
document type, TDocument.  This provides cleaner support for generic type
detection in the new extension methods.

Extended LookupInBuilder<T>, MutateInBuilder<T>, and DocumentFragment<T>
with the new ITypeSerializerProvider interface.  This allows them to
expose the type serializer, which is necessary for lambda path evaluation,
without confusing their core interfaces.  Also, the original builder used
to build the request is passed to the resulting DocumentFragment<T> so
that it uses the same serializer for path evaluation.

Created a system for evaluating lambda expression paths and converting
them to path strings in the Couchbase.Core.ExpressionVisitors namespace.
These are currently internal so they can be refactored as needed without
breaking any exposed API.

Added extension methods in SubdocExtensions that add path evaluation
functions for all applicable methods in ILookupInBuilder<T>,
IMutateInBuilder<T>, and IDocumentFragment<T>.

Created unit tests for the lambda path evaluation process and for the
extension methods.

Results
-------
Lambda paths can be easily built inline.  This reduces the chance for
semantic errors that don't appear until the application is deployed, since
the compiler can now do some validation.  It also adds support for
intellisense while building the path.  All applicable operations that take
a path are supported.

Some of the operations are technically somewhat more limited than their
plain string counterparts due to type restrictions.  For example, you
can't replace a string property with a JSON object, unless the type of the
property was object or dynamic.  However, this seems preferable since it
reduces the likelihood of error.  If the property is meant to be more
flexible then it would have an object or dynamic type.  Also, manually
built string paths can still be used for these advanced use cases.

Support for using dynamic types is somewhat limited.  Dynamics can only be
used at the final leaf of the subdocument path.  They can't be used as the
root document, or in any intermediate nodes of the path.  This is due to
limitations in .Net related to lambda expressions.  Again, manually built
string paths can still be used for these advanced use cases.

Also note that trying to pass dynamic fragments into mutation commands
will require that they be typecast to object or some othere non-dynamic
type first.  This is because extension methods can't be called with
dynamic parameters.

Change-Id: I3a5e26f59cb0a18653345fd5abed786b21b7a65d
Reviewed-on: http://review.couchbase.org/60975
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>