NCBC-1095: Do not retry on NodeUnavailableException
Motivation
----------
If an operation fails while the client is transitioning into a "down node"
state, an operation may continue to retry instead of returning wit
NodeUnavailable.

Modifications
-------------
Make OperationResult.ShouldRetry return false if status is
NodeUnavailable.

Results
-------
An operation will only retry if it's OperationResult is NMV and in some
cases only if it has a CAS value. One other state is a ClientFailure
meaning that the operation never was sent over the wire.

Change-Id: Ibba75a5896a1658a833aae2acd1340f50fed17a5
Reviewed-on: http://review.couchbase.org/61546
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>