NCBC-1098: Use HttpClient for N1QL queries
Motivation
----------
WebHttpRequest is not available in .Net Core, so refactoring to use
System.Net.Http is required to support .Net Core compilation.

Modifications
-------------
Updated QueryClient.ExecuteQuery to internally use
QueryClient.ExecuteQueryAsync, blocking the thread to wait for the
asynchronous task to complete.

Results
-------
WebHttpRequest is no longer used by N1QL queries.  It also does not appear
to have a significant impact on performance.

When wrapping synchronously waiting on asynchronous code, there is a risk
of deadlocks.  This is mitigated by temporarily removing the
SynchronizationContext from the thread before starting the task.  It is
still possible that it might deadlock if the ThreadPool runs out of
avaiable threads, but this is unlikely.

Change-Id: I869b3147b94f9357ea1495d3b7c7996f871c5951
Reviewed-on: http://review.couchbase.org/62339
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>