NCBC-1061: Improve vBucket Retry Logic
Motivation
----------
Make the .NET Not My VBucket (NMV) behavior consistent with others SDK's
per RFC:0005_vbucket_retries.

Modifications
-------------
Make overload for GetServer take a cluster map rev # which is that latest
revision that the operation was executed against that uses the VBucket
forward map if it's available on a retry, otherwise uses the regular
VBucket Map. Change the rev# on various classes to be a unsigned integer
since a rev # should always be a positive integer. Update various code
paths so that the rev# is included with each operation retry.

Results
-------
Client will use the VBucket Forward Map if it's available if a NMV is
returned by the server, but only if the rev # hasn't changed, otherwise it
will use the regular VBucket Map.

Change-Id: I47d1806092ecce0ab59cf07ed614380d7c8bba6a
Reviewed-on: http://review.couchbase.org/62881
Reviewed-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>