NCBC-1106: Prevent null argument exception in Connection.Send
Motivation
----------
Occasionally in production operation, an ArgumentNullException is seen
within Connection.Send, when the dst parameter of BlockCopy is null.

Modifications
-------------
If the BufferAllocator is out of buffers when a Connection is
instantiated, SetBuffer returns false.  In this scenario, throw an
immediate exception which will be logged by the ConnectionPool
connection factory.

To help prevent the leak in the first place, ensure that the buffer
is released in Dispose, even if other actions fail to complete.

Results
-------
Theoretically, the BufferAllocator should never run out of buffers because
its buffer size is equal to the maximum size of the connection pool.
However, as a practical matter it is possible for other errors to cause
the buffer pool to be exhausted.

These changes ensure that there is better logging in place if this
happens, which will make it easier to diagnose in the future.
Additionally, one possible hole has been filled to help prevent the
situation from occuring in the first place.  While the underlying problem
may not be fixed, if it does occur in the future it should be somewhat
easier to identify and track down.

Change-Id: If90b8dc4f10f0424a78cc39387b45012bfef3789
Reviewed-on: http://review.couchbase.org/63214
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>