NCBC-1098: Use HttpClient for view queries
Motivation
----------
WebHttpRequest is not available in .Net Core, so refactoring to use
System.Net.Http is required to support .Net Core compilation.

Modifications
-------------
Updated ViewClient.Execute to internally use ViewClient.ExecuteAsync,
blocking the thread to wait for the asynchronous task to complete.

Since bad URI requests no longer return a WebException, change the related
tests to ensure that any exception is acceptable.  Note that the specific
timeout exception in this test returns a TaskCanceledException, and the
Error property contains "The request has timed out."

Results
-------
WebHttpRequest is no longer used by view queries.  It also does not appear
to have a significant impact on performance.

When wrapping synchronously waiting on asynchronous code, there is a risk
of deadlocks.  This is mitigated by temporarily removing the
SynchronizationContext from the thread before starting the task.  It is
still possible that it might deadlock if the ThreadPool runs out of
available threads, but this is unlikely.

Change-Id: I848ab46edb02fd76cb539e571d29c9909070b8a3
Reviewed-on: http://review.couchbase.org/63217
Reviewed-by: Jeffry Morris <jeffrymorris@gmail.com>
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>