NCBC-1110: GetFromReplica times out if key does not exist
Motivation
----------
If an operation results in a KeyNotFound status, it should bypass the
retry process and return immediately.

Modifications
-------------
Add an additional check to ensure that the status is retryable -
KeyNotFound is not retryable so the request will return immediatly as
failed and a status of KeyNotFound.

Results
-------
KeyNotFound will not cause a retry.

Change-Id: I45152ee6bd8558a841e798081cc03e4347dea744
Reviewed-on: http://review.couchbase.org/63776
Tested-by: Jeffry Morris <jeffrymorris@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>